fix: 취침 알림 시간 설정 API 버그를 수정한다
* fix: if문으로 null이 아닐 때만 delete 메소드를 실행하도록 수정

+++ src/main/java/go/alarm/service/UserServiceImpl.java
@@ -32,7 +32,10 @@ public class UserServiceImpl implements UserService {
     @Override
     public User setBedTime(Long userId, UserRequestDTO.SetBedTimeDTO request) {
         User user = userRepository.findById(userId).get();
-        dayOfWeekRepository.delete(user.getBedDayOfWeek());
+
+        if(user.getBedDayOfWeek() != null){
+            dayOfWeekRepository.delete(user.getBedDayOfWeek());
+        }
 
         DayOfWeek dayOfWeek = dayOfWeekConverter.toDayOfWeek(request.getBedDayOfWeekList());
